requirements.txt
django
django-cors-headers
djangorestframework
markdown
django-filter


googletrans
google_trans_new
deep-translator

Pillow
psycopg2
djangorestframework-simplejwt

free-proxy
#g4f
#gradio_client

supabase